StoredProcedureQuery getSingleResultOrNull() fix
