+++ jpa/eclipselink.jpa.testapps/jpa.test.advanced2/src/test/java/org/eclipse/persistence/testing/tests/advanced2/CriteriaQueryTest.java
@@ -15,7 +15,7 @@
 //       - 367452: JPA 2.1 Specification support for joins with ON clause
 //     09/26/2012-2.5 Chris Delahunt
 //       - 350469: JPA 2.1 Criteria Query framework Bulk Update/Delete support
-package org.eclipse.persistence.testing.tests.jpa21.advanced;
+package org.eclipse.persistence.testing.tests.advanced2;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Query;
