+++ jpa/eclipselink.jpa.testapps/jpa.test.advanced2/src/test/java/org/eclipse/persistence/testing/tests/advanced2/ddl/DDLSchemaTest.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2012, 2022 Oracle and/or its affiliates. All rights reserved.
- * Copyright (c) 2012, 2022 IBM Corporation. All rights reserved.
+ * Copyright (c) 2012, 2024 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2024 IBM Corporation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -30,7 +30,7 @@
 //       - 389090: JPA 2.1 DDL Generation Support
 //     02/02/2015-2.6.0 Dalia Abo Sheasha
 //       - 458462: generateSchema throws a ClassCastException within a container
-package org.eclipse.persistence.testing.tests.jpa21.advanced.ddl;
+package org.eclipse.persistence.testing.tests.advanced2.ddl;
 
 import jakarta.persistence.Persistence;
 import junit.framework.Test;
