+++ jpa/eclipselink.jpa.testapps/jpa.test.advanced2/src/test/java/org/eclipse/persistence/testing/tests/advanced2/ddl/ForeignKeyTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2022 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -21,7 +21,7 @@
 //       - 350487: JPA 2.1 Specification defined support for Stored Procedure Calls
 //     04/12/2013-2.5 Guy Pelletier
 //       - 405640: JPA 2.1 schema generation drop operation fails to include dropping defaulted fk constraints.
-package org.eclipse.persistence.testing.tests.jpa21.advanced.ddl;
+package org.eclipse.persistence.testing.tests.advanced2.ddl;
 
 import jakarta.persistence.EntityManager;
 import junit.framework.Test;
